STRM-140 bumped to use pravega master, updated package name, updated flink to 1.3-SNAPSHOT
