+++ anomaly-detection/src/main/java/io/pravega/anomalydetection/event/pipeline/IPipeline.java
@@ -1,0 +1,12 @@
+/**
+ *
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries.
+ *
+ */
+package io.pravega.anomalydetection.event.pipeline;
+
+import io.pravega.anomalydetection.event.AppConfiguration;
+
+public interface IPipeline {
+	void run(AppConfiguration appConfiguration) throws Exception;
+}
