+++ flink-examples/build.gradle
@@ -12,15 +12,15 @@ archivesBaseName = 'pravega-flink-examples'
 
 dependencies {
     compile "org.scala-lang.modules:scala-java8-compat_2.11:0.3.0"
-    compile "io.pravega:pravega-connectors-flink_2.11:${pravegaVersion}"
-    compileOnly group: 'org.apache.flink', name: 'flink-streaming-java_2.11', version: '1.2.0'
-    compileOnly group: 'org.apache.flink', name: 'flink-streaming-scala_2.11', version: '1.2.0'
+    compile ("io.pravega:pravega-connectors-flink_2.11:${pravegaVersion}:all")
+    compileOnly ("org.apache.flink:flink-streaming-java_2.11:${flinkVersion}")
+    compileOnly ("org.apache.flink:flink-streaming-scala_2.11:${flinkVersion}")
 }
 
 shadowJar {
     dependencies {
         include dependency("org.scala-lang.modules:scala-java8-compat_2.11")
-        include dependency("io.pravega:pravega-connectors-flink_2.11")
+        include dependency("io.pravega:pravega-connectors-flink_2.11:${pravegaVersion}:all")
     }
 }
