+++ flink-examples/src/main/scala/io/pravega/examples/flink/util/PravegaParameters.scala
