+++ flink-examples/src/main/scala/io/pravega/examples/flink/util/serialization/PravegaSerializationSchema.java
