+++ gradle.properties
@@ -8,3 +8,5 @@ pravegaVersion=0.1.0-SNAPSHOT
 ### outputs
 samplesVersion=0.1.0-SNAPSHOT
 
+### Flink
+flinkVersion=1.3-SNAPSHOT
